Address audit findings 3, 5, and 11 (#1131)
* refactor: add blockTimestamp param in VestingMath functions

refactor: remove unchecked in calculateStreamedPercentage
docs: add assumptions in VestingMath function
docs: improve explanatory comments

* docs: fix typos

* docs: polish natspec

* refactor: move logic in VestingMath

chore: remove MAX_COUNT assumption

* docs: polish assumptions

* refactor: use unchecked and casting in calculateStreamedPercentage

docs: last polishes

---------

+++ src/LockupNFTDescriptor.sol
@@ -200,7 +200,7 @@ contract LockupNFTDescriptor is ILockupNFTDescriptor {
     {
         // This cannot overflow because both inputs are uint128s, and zero deposit amounts are not allowed in Sablier.
         unchecked {
-            return streamedAmount * 10_000 / depositedAmount;
+            return uint256(streamedAmount) * 10_000 / depositedAmount;
         }
     }
